Fix autodetection of NVIDIA GPUs
We need to remove the `[ ... ]` when testing the return value of
calling the is_nvidia_gpu_installed() method, otherwise we will
be testing only if the function is run, regardless of its actual
return value, and we will always get a "true" result.

This should fix the autodetection of NVIDIA GPUs on NVIDIA-less
systems, where wkdev-setup-nvidia-gpu-for-container would be always
run otherwise.