wkdev-create: Prefer hostnamectl instead of hostname
Prefer using hostnamectl instead of the hostname program. The
hostnamectl program is provided along systemd, which is already
a dependency of wkdev-sdk, while the hostname program is optional
for a modern Linux system and might not be available.

This prevents the following error in systems which do not have the
"hostname" program installed:

  .../scripts/host-only/wkdev-create: line 67: hostname: command not found