Handle WAYLAND_DISPLAY changing between runs
Wayland is unique here in that it places its sockets in $XDG_RUNTIME_DIR directly.

This means that when it changes we need to ensure $XDG_RUNTIME_DIR/$WAYLAND_DISPLAY
is in the containers runtime directory.

Other variables like $DISPLAY are fine because they are in a subdirectory we already
have mounted.