wkdev-init: Only run init tasks on creation
Previously it would re-run the tasks every time. I'm not sure this was intended
and all of the tasks it currently does seem to be ones we only want once.

This caused issues when an old container was started as the init script made
assumptions about the container being a newer version and would fail.