wkdev-create: Set pull=newer option explicitly
Fix the expected logic of the command by ensuring the local image
used for creation is up-to-date with the one in the remote
repository.

With --pull=newer the podman pull command will pull the image on the
registry if this is newer than the one in the local containers
storage BUT an image is considered to be newer when the digests are
different because the comparing the time stamps is prone to errors.