jhbuild: Add libspiel
Spiel is a nicer alternative to Festival for speech synthesys because its API is directly inspired
from WebSpeech. A future patch will rewrite our speech synthesys backend using Spiel, but we first
need it in the SDK. The host is responsible for installing voice providers using Flatpak, so this
part is not included in this patch.