wkdev-enter: Show container log, while waiting for initialization (#57)
Show an excerpt from the container log after waiting for N seconds to
try to enter the container.

This helps the user to indicate progress, when entering a freshly
created container for the first time (when wkdev-init is running for the
first time). At the same time, no output is shown upon successive
wkdev-enter commands, since the container is already initialized, and
entering it is fast by definition.