Fix setting environment vars in GitHub workflow.
GITHUB_ENV needs to be used to set environment variables,
not GITHUB_OUTPUT. The key/value pairs are also propagated
to the following steps, so it works.

Now that the environment variable setting works, fix a problem
with "WKDEV_SDK_CONTAINER_REGISTRY_USER_NAME" -- the user name
needs to be in lower case (aka. 'igalia', not 'Igalia').