chown the directories in the image and skip chown of the first entry (#184)
~~~
Previously, it did chown /opt/rust and /jhbuild directories to the
user's UID
at the first time of wkdev-enter. It took a while to finish.

In the most cases, the user's UID is 1000. Added a fast path by assuming
the
UID is 1000. Execute chown for the directories in Containerfile. And,
skip the
chown at the first entry if the directories are already writable.
~~~